// src/content/generated/youtube.ts
export interface YtVideo {
  id: string;
  title: string;
  thumb: string;
  seconds: number;
  short: boolean;
}